% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab_utils.R
\name{check_gitlab_token}
\alias{check_gitlab_token}
\title{Check GitLab token and open .Renviron if needed}
\usage{
check_gitlab_token()
}
\value{
TRUE if token exists, otherwise FALSE and shows instructions
}
\description{
Check GitLab token and open .Renviron if needed
}
