% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creator.R
\name{create_analysis_template}
\alias{create_analysis_template}
\title{Create a full analysis template}
\usage{
create_analysis_template(
  path,
  project_name = "analysis_project",
  git = TRUE,
  gitlab = TRUE,
  gitlab_url = "https://git.bihealth.org",
  open = TRUE
)
}
\arguments{
\item{path}{Project path}

\item{project_name}{Name of the project}

\item{gitlab_url}{GitLab URL (default: git.bihealth.org)}
}
\description{
Create a full analysis template
}
