% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R, R/addin_ui.R
\name{create_analysis_template_addin}
\alias{create_analysis_template_addin}
\title{Interaktives Addin zur Erstellung eines Analyseprojekts}
\usage{
create_analysis_template_addin()

create_analysis_template_addin()
}
\description{
Interaktives Addin zur Erstellung eines Analyseprojekts

RStudio Addin UI zur Erstellung eines Analyseprojekts
}
