% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab_utils.R, R/helper_gitlab.R
\name{create_gitlab_project}
\alias{create_gitlab_project}
\title{Create a GitLab project using the GitLab API}
\usage{
create_gitlab_project(
  project_name,
  visibility = "private",
  gitlab_url = "https://git.bihealth.org"
)

create_gitlab_project(
  project_name,
  visibility = "private",
  gitlab_url = "https://git.bihealth.org"
)
}
\arguments{
\item{project_name}{Name of the new GitLab project}

\item{visibility}{"private", "internal", or "public"}

\item{gitlab_url}{URL to GitLab server}
}
\value{
Repository URL (ssh or https)

Repository SSH URL if successful, otherwise NULL
}
\description{
Create a GitLab project using the GitLab API

Create GitLab project via API and return repository URL
}
