\name{EpiR}
\alias{EpiR}
\title{EpiR}
\description{Biblioteca e interface gráfica para análises epidemiológicas em \R
}
\usage{
EpiR(gui = TRUE)
}
\arguments{
    \item{gui}{valor lógico indicando que a interface gráfica deve ser carregada}
}
\value{
Esta função não retorna valor.
}
\author{
Equipe Epi-R \url{http://www.ims.uerj.br/epi-r}
}



