% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_gitlab.R
\name{helper_check_and_abort}
\alias{helper_check_and_abort}
\title{Check for GitLab token and abort setup with instructions if missing}
\usage{
helper_check_and_abort(gitlab_url = "https://git.bihealth.org")
}
\arguments{
\item{gitlab_url}{Character. URL to the GitLab server. Default: "https://git.bihealth.org"}
}
\value{
TRUE if token is present; otherwise aborts with message and returns NULL
}
\description{
Check for GitLab token and abort setup with instructions if missing
}
