% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creator.R
\name{create_analysis_template}
\alias{create_analysis_template}
\title{Create a complete analysis project folder with optional Git and GitLab integration}
\usage{
create_analysis_template(path, project_name = "analysis_project")
}
\arguments{
\item{path}{Character. Path to base directory.}

\item{project_name}{Character. Project folder name.}
}
\description{
Create a complete analysis project folder with optional Git and GitLab integration
}
