% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_creator.R
\name{create_analysis_template}
\alias{create_analysis_template}
\title{Create a project template with folders and Quarto file}
\usage{
create_analysis_template(path, project_name = "analysis_project")
}
\arguments{
\item{path}{Character. Path where the analysis project will be created.}

\item{project_name}{Character. Name for the project folder and Rproj file.}
}
\description{
Create a project template with folders and Quarto file
}
