% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab_utils.R
\name{create_gitlab_project}
\alias{create_gitlab_project}
\title{Create a GitLab project using the GitLab API}
\usage{
create_gitlab_project(project_name, visibility = "private")
}
\arguments{
\item{project_name}{Name of the project to create}

\item{visibility}{"private" (default), "internal", or "public"}
}
\value{
Repository URL (ssh or https)
}
\description{
Create a GitLab project using the GitLab API
}
