\name{epi.about}

\alias{epi.about}

\title{The library epiR: summary information}

\description{
An R package for the analysis of epidemiological data.
}

\usage{
epi.about()
}

\author{
Mark Stevenson (\email{mark.stevenson1@unimelb.edu.au}), Faculty of Veterinary Science, University of Melbourne, Parkville, Victoria 3052, Australia.

Simon Firestone, Faculty of Veterinary Science, University of Melbourne, Parkville, Victoria 3052, Australia.

Telmo Nunes, UISEE/DETSA, Faculdade de Medicina Veterinaria --- UTL, Rua Prof. Cid dos Santos, 1300 - 477 Lisboa Portugal.

Javier Sanchez, Atlantic Veterinary College, University of Prince Edward Island, Charlottetown, Prince Edward Island, C1A 4P3, Canada.

Ron Thornton, Ministry for Primary Industries New Zealand, PO Box 2526 Wellington, New Zealand.
}

\keyword{univar}
