\name{epi.about}

\alias{epi.about}

\title{The library epiR: summary information}

\description{
Functions for quantitative epidemiology.
}

\usage{
epi.about()
}

\details{
The most recent version of the \code{epiR} package can be obtained from: \url{http://epicentre.massey.ac.nz/}
}

\author{
Mark Stevenson, EpiCentre, IVABS, Massey University, Palmerston North New Zealand.

Telmo Nunes, UISEE/DETSA, Faculdade de Medicina Veterinaria --- UTL, Rua Prof. Cid dos Santos, 1300 - 477 Lisboa Portugal.

Javier Sanchez, Atlantic Veterinary College, University of Prince Edward Island, Charlottetown, Prince Edward Island, C1A 4P3, Canada.

Ron Thornton, MAF New Zealand, PO Box 2526 Wellington, New Zealand.
}

\keyword{univar}
